make it DOS compatible :^)

+++ src/timer.h
@@ -6,7 +6,7 @@
 // number chosen to be integer divisor of PIC frequency
 #define TIMER_TPS 363
 
-u64 timer_get();
+u32 timer_get();
 void timer_init();
 
 #endif
